update `complexobject.rs` for 3.13
